numeryczne-cwiczenia-10

+++ metody-numeryczne/cwiczenia-9/LCG_generator.h
@@ -4,11 +4,14 @@
 class LCG_generator
 {
 public:
-	LCG_generator(int _seed, int _mult = 869, int _add = 19, int _mod = 1318492) : mult(_mult), add(_add), mod(_mod), seed(_seed) { x = seed; }
-	int random();
+	LCG_generator(unsigned int _seed,
+		unsigned int _mult = 869,
+		unsigned int _add = 19,
+		unsigned int _mod = 1318492) : mult(_mult), add(_add), mod(_mod), seed(_seed) { x = seed; }
+	double random();
 	void reset();
 private:
-	int mult, add, mod, seed, x;
+	unsigned int mult, add, mod, seed, x;
 };
 
 #endif // LCG_GENERATOR
